requirements.txt
# requirements.txt
# Inferred dependencies for the VietSGG project.
# Please review and pin versions as appropriate for your environment.
kagglehub

numpy>=1.24
opencv-python>=4.7.0
Pillow>=9.0.0
matplotlib>=3.6.0
pandas>=1.5.0

tqdm>=4.64.0
transformers
